pyproject.toml
[project]
name = "stock-filter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "claude-agent-sdk>=0.1.56",
    "fastapi>=0.135.3",
    "fastmcp>=3.2.0",
    "google-genai>=1.70.0",
    "openai-agents>=0.13.5",
]
